% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_name.R
\name{cat_name}
\alias{cat_name}
\title{Get cats by name}
\usage{
cat_name(df, text = "nut")
}
\arguments{
\item{df}{A dataset with column name}

\item{text}{Partial name of a cat}
}
\value{
The names of cats in descending popularity
}
\description{
Get cats by name
}
